requirements.txt
setuptools=57.0.0
requests==2.25.1
python-dotenv==0.15.0
urllib3==1.26.3